iox-#1032 Move error checking functionality to new header
